// web/src/api/generated/models/book-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface BookDTO {
    'id'?: number;
    'title'?: string;
    'author'?: string;
    'coverUrl'?: string;
    'fileType'?: string;
    'status'?: string;
    'totalChapters'?: number;
    'wordCount'?: number;
    'fileSize'?: number;
    'createTime'?: string;
    'updateTime'?: string;
}